Delete endpoints
Added a deleteTvShowReminder endpoint and the service logic for this operation. Also i added a trigger in the database that when we delete a reminder, if this reminder has a user tv show associated it, we deleted it. Also i added a deleteUserTvShow endpoint, with delete on cascade in the DB,  we also delete the reminder associated with it

+++ src/main/java/com/example/spring/services/TvShowReminderService.java
@@ -189,8 +189,10 @@ public class TvShowReminderService {
     }
 
     // Delete a tv show reminder.
-    public void deleteTvShowReminder(Integer idTvShowReminder) {
-        tvShowReminderRepository.deleteById(idTvShowReminder);
+    public void deleteTvShowReminder(Integer idUser, Integer idTvShowReminder) throws ResourceNotFoundException {
+        if(getTvShowReminder(idUser,idTvShowReminder) != null){
+            tvShowReminderRepository.deleteById(idTvShowReminder);
+        }
     }
 
     // Done
